requirements.txt
ConferenceCorpus
tabulate
pylodstorage
openpyxl
geograpy3
nominatim
OSMPythonTools==0.3.0
spacy
dataclasses
requests

lxml
bs4